requirements.txt
# Core Streamlit and Web Framework
streamlit>=1.28.0
fastapi>=0.100.0
uvicorn>=0.23.0

# Free AI/ML Libraries - No paid APIs
google-generativeai>=0.3.0
transformers>=4.30.0
torch>=2.0.0
spacy>=3.6.0
langchain>=0.1.0
langchain-google-genai>=1.0.0
langgraph>=0.0.40

# Document Processing - Free libraries
python-docx>=0.8.11
PyPDF2>=3.0.1
pypdf>=3.0.0
pdfplumber>=0.9.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
typing-extensions>=4.5.0

# Logging
loguru>=0.7.0

# File handling
chardet>=5.1.0
